// planning-idsdp-pps-frontend/src/app/main/modules/rdpp-rtapp-management/models/fiscal-year-wise-total-cost.model.ts
export class FiscalYearWiseTotalCost {
  fiscalYear?: string = '';
  qty?: number = 0;
  totalAmount: number = 0;
  gobAmount: number = 0;
  gobFeAmount: number = 0;
  gobThruAmount: number = 0;
  spAcAmount: number = 0;
  thruPdAmount: number = 0;
  thruDpAmount: number = 0;
  ownFundAmount: number = 0;
  ownFundFeAmount: number = 0;
  otherAmount: number = 0;
  otherFeAmount: number = 0;
}